% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetrieveAnswerFunction.R
\name{Retrieve_answer}
\alias{Retrieve_answer}
\title{Deliverable1}
\usage{
Retrieve_answer(option)
}
\arguments{
\item{option}{(Allows to select any of the marked exercises from 1 to 6)}
}
\value{
The solution of the selected exercise
}
\description{
Function to choose any of the exercises previously done in class
}
\examples{
Retrieve_answer(1)
Retrieve_answer(5)
}
